# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YourApp API",
    "description": "Generate beautiful PDFs from HTML/CSS templates with Jinja2 via API. Design once, generate thousands.",
    "version": "1.4.1.dev16+g7eb0207e9.d20260215"
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/v1/account": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Get account info",
        "description": "Get account information including remaining credits.\n\n**Authentication:** API Key required (`x-api-key` header) or JWT token\n\n**Usage:** Check credit balance before performing operations.\n\n**No credits consumed:** This is a read-only endpoint.\n\n**Rate Limits:** 60 requests/min (free), 120 requests/min (paid). Headers included in response.",
        "operationId": "getAccount",
        "responses": {
          "200": {
            "description": "Account information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoResponse"
                },
                "example": {
                  "credits": 150,
                  "email": "user@example.com"
                }
              }
            }
          },
          "403": {
            "description": "Invalid or missing API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/account/transactions": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "List transactions",
        "description": "List transaction history for the authenticated user.\n\n**Authentication:** API Key required (`x-api-key` header) or JWT token\n\n**Pagination:** Use `limit` and `offset` query parameters.\n\n**Transaction types:**\n- `PDFGEN`: PDF generation (consumes credits)\n- `REFUND`: Credit refund (on failed generation)\n- `PURCHASE`: Credit purchase\n- `BONUS`: Bonus credits\n\n**Credits field:**\n- Positive value = credits consumed\n- Negative value = credits added\n\n**No credits consumed:** This is a read-only endpoint.\n\n**Rate Limits:** 60 requests/min (free), 120 requests/min (paid). Headers included in response.",
        "operationId": "listTransactions",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of records to return",
              "default": 300,
              "title": "Limit"
            },
            "description": "Number of records to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of records to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                },
                "example": {
                  "transactions": [
                    {
                      "transaction_ref": "550e8400-e29b-41d4-a716-446655440000",
                      "transaction_type": "PDFGEN",
                      "template_id": "HMQywVpZxqAM",
                      "exec_tm": 1250,
                      "credits": 1,
                      "created_at": "2024-01-20T14:30:00Z"
                    },
                    {
                      "transaction_ref": "660e8400-e29b-41d4-a716-446655440001",
                      "transaction_type": "PURCHASE",
                      "credits": -100,
                      "created_at": "2024-01-15T10:00:00Z"
                    }
                  ],
                  "total": 42,
                  "limit": 300,
                  "offset": 0
                }
              }
            }
          },
          "403": {
            "description": "Invalid or missing API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountInfoResponse": {
        "properties": {
          "credits": {
            "type": "integer",
            "title": "Credits",
            "description": "Remaining credits",
            "examples": [
              150
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "User email address",
            "examples": [
              "user@example.com"
            ]
          }
        },
        "type": "object",
        "required": [
          "credits"
        ],
        "title": "AccountInfoResponse",
        "description": "Response for account info endpoint",
        "examples": [
          {
            "credits": 150,
            "email": "user@example.com"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Transaction": {
        "properties": {
          "transaction_ref": {
            "type": "string",
            "title": "Transaction Ref",
            "description": "Unique transaction reference (UUID)",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          },
          "transaction_type": {
            "type": "string",
            "title": "Transaction Type",
            "description": "Transaction type: PDFGEN, PURCHASE, REFUND, BONUS",
            "examples": [
              "PDFGEN"
            ]
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id",
            "description": "Template short ID (for PDFGEN)",
            "examples": [
              "HMQywVpZxqAM"
            ]
          },
          "exec_tm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exec Tm",
            "description": "Execution time in milliseconds",
            "examples": [
              1250
            ]
          },
          "credits": {
            "type": "integer",
            "title": "Credits",
            "description": "Credits consumed (positive) or added (negative)",
            "examples": [
              1
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO 8601 timestamp",
            "examples": [
              "2024-01-20T14:30:00Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "transaction_ref",
          "transaction_type",
          "credits",
          "created_at"
        ],
        "title": "Transaction",
        "description": "Transaction record"
      },
      "TransactionsResponse": {
        "properties": {
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "type": "array",
            "title": "Transactions"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of transactions",
            "examples": [
              42
            ]
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Number of records returned",
            "examples": [
              300
            ]
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Number of records skipped",
            "examples": [
              0
            ]
          }
        },
        "type": "object",
        "required": [
          "transactions",
          "total",
          "limit",
          "offset"
        ],
        "title": "TransactionsResponse",
        "description": "Response for transactions list endpoint",
        "examples": [
          {
            "limit": 300,
            "offset": 0,
            "total": 42,
            "transactions": [
              {
                "created_at": "2024-01-20T14:30:00Z",
                "credits": 1,
                "exec_tm": 1250,
                "template_id": "HMQywVpZxqAM",
                "transaction_ref": "550e8400-e29b-41d4-a716-446655440000",
                "transaction_type": "PDFGEN"
              },
              {
                "created_at": "2024-01-15T10:00:00Z",
                "credits": -100,
                "transaction_ref": "660e8400-e29b-41d4-a716-446655440001",
                "transaction_type": "PURCHASE"
              }
            ]
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication. Get yours at https://example.com/dashboard/api-keys"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}